Review remaining commands
